Re-structure Chrome_Test.csproj
