+++ Web/Chrome_Test/Chrome_Test.csproj
@@ -132,68 +132,68 @@
       <RepositoryCode>true</RepositoryCode>
       <id>7d0be441-e59f-4caa-8815-162e4b68c5f8</id>
     </Compile>
-    <Compile Include="CloseBrowser.cs">
-      <BoundTo>f2873c10-8693-42cd-a0c8-d76af30a8b60</BoundTo>
-      <DependentUpon>CloseBrowser.rxrec</DependentUpon>
-      <RecorderCode>true</RecorderCode>
-      <id>9d9281d3-79ce-497c-9f59-cc345cb0ff4b</id>
-    </Compile>
-    <Compile Include="CloseBrowser.UserCode.cs">
-      <BoundTo>f2873c10-8693-42cd-a0c8-d76af30a8b60</BoundTo>
-      <DependentUpon>CloseBrowser.rxrec</DependentUpon>
-      <UserCode>true</UserCode>
-      <id>450384f5-69eb-4372-a18f-502554dcb72f</id>
-    </Compile>
-    <Compile Include="BackToHomePage.cs">
-      <BoundTo>43185cca-9174-42ce-bf52-e4e97de544ea</BoundTo>
+    <Compile Include="Recordings\BackToHomePage.cs">
       <DependentUpon>BackToHomePage.rxrec</DependentUpon>
       <RecorderCode>true</RecorderCode>
       <id>2febc236-fe83-4952-9b75-946a37b52c68</id>
-    </Compile>
-    <Compile Include="BackToHomePage.UserCode.cs">
       <BoundTo>43185cca-9174-42ce-bf52-e4e97de544ea</BoundTo>
+    </Compile>
+    <Compile Include="Recordings\BackToHomePage.UserCode.cs">
       <DependentUpon>BackToHomePage.rxrec</DependentUpon>
       <UserCode>true</UserCode>
       <id>0d7d6022-df36-4b86-a7aa-b9d30c57fb3c</id>
+      <BoundTo>43185cca-9174-42ce-bf52-e4e97de544ea</BoundTo>
     </Compile>
-    <Compile Include="Search.cs">
-      <BoundTo>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</BoundTo>
-      <DependentUpon>Search.rxrec</DependentUpon>
+    <Compile Include="Recordings\CloseBrowser.cs">
+      <DependentUpon>CloseBrowser.rxrec</DependentUpon>
       <RecorderCode>true</RecorderCode>
-      <id>3e067bed-4a84-4692-ab3c-b97fe4db74ab</id>
+      <id>9d9281d3-79ce-497c-9f59-cc345cb0ff4b</id>
+      <BoundTo>f2873c10-8693-42cd-a0c8-d76af30a8b60</BoundTo>
     </Compile>
-    <Compile Include="Search.UserCode.cs">
-      <BoundTo>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</BoundTo>
-      <DependentUpon>Search.rxrec</DependentUpon>
+    <Compile Include="Recordings\CloseBrowser.UserCode.cs">
+      <DependentUpon>CloseBrowser.rxrec</DependentUpon>
       <UserCode>true</UserCode>
-      <id>174e3976-c0bd-45a2-a71d-69186a35d3cd</id>
+      <id>450384f5-69eb-4372-a18f-502554dcb72f</id>
+      <BoundTo>f2873c10-8693-42cd-a0c8-d76af30a8b60</BoundTo>
     </Compile>
-    <Compile Include="OpenBrowser.cs">
-      <BoundTo>716e1a36-fbdd-4a8b-b18e-6b0190f3c31b</BoundTo>
+    <Compile Include="Recordings\OpenBrowser.cs">
       <DependentUpon>OpenBrowser.rxrec</DependentUpon>
       <RecorderCode>true</RecorderCode>
       <id>59986f2b-fd23-4e60-a056-7e13568c0f69</id>
-    </Compile>
-    <Compile Include="OpenBrowser.UserCode.cs">
       <BoundTo>716e1a36-fbdd-4a8b-b18e-6b0190f3c31b</BoundTo>
+    </Compile>
+    <Compile Include="Recordings\OpenBrowser.UserCode.cs">
       <DependentUpon>OpenBrowser.rxrec</DependentUpon>
       <UserCode>true</UserCode>
       <id>057dd530-fea3-4ce8-b52e-36cf18dd09b6</id>
+      <BoundTo>716e1a36-fbdd-4a8b-b18e-6b0190f3c31b</BoundTo>
     </Compile>
     <Compile Include="Program.cs">
       <id>29e4f2e8-51b9-408a-902c-5a31e0f46a17</id>
     </Compile>
-    <Compile Include="Validation.cs">
-      <BoundTo>02e26d06-ea6d-4fad-8e56-9ad1339c7022</BoundTo>
+    <Compile Include="Recordings\Search.cs">
+      <DependentUpon>Search.rxrec</DependentUpon>
+      <RecorderCode>true</RecorderCode>
+      <id>3e067bed-4a84-4692-ab3c-b97fe4db74ab</id>
+      <BoundTo>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</BoundTo>
+    </Compile>
+    <Compile Include="Recordings\Search.UserCode.cs">
+      <DependentUpon>Search.rxrec</DependentUpon>
+      <UserCode>true</UserCode>
+      <id>174e3976-c0bd-45a2-a71d-69186a35d3cd</id>
+      <BoundTo>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</BoundTo>
+    </Compile>
+    <Compile Include="Recordings\Validation.cs">
       <DependentUpon>Validation.rxrec</DependentUpon>
       <RecorderCode>true</RecorderCode>
       <id>e4c3daaf-0faf-454d-a0cc-63f8498c81d8</id>
-    </Compile>
-    <Compile Include="Validation.UserCode.cs">
       <BoundTo>02e26d06-ea6d-4fad-8e56-9ad1339c7022</BoundTo>
+    </Compile>
+    <Compile Include="Recordings\Validation.UserCode.cs">
       <DependentUpon>Validation.rxrec</DependentUpon>
       <UserCode>true</UserCode>
       <id>7fef0b7a-4926-43b0-a85c-ca8474d5e3ba</id>
+      <BoundTo>02e26d06-ea6d-4fad-8e56-9ad1339c7022</BoundTo>
     </Compile>
   </ItemGroup>
   <ItemGroup>
@@ -211,27 +211,27 @@
       <repoid>49cf3a26-a542-4035-b0ff-0e29ce9bcbe4</repoid>
       <reporootid>b95a0521-1a6a-49fe-a98f-206201c64030</reporootid>
     </None>
-    <None Include="CloseBrowser.rxrec">
-      <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
-      <id>f2873c10-8693-42cd-a0c8-d76af30a8b60</id>
-      <recordtableid>3407bd7f-4b1b-46f5-a4e7-1c2a187f854f</recordtableid>
-    </None>
-    <None Include="BackToHomePage.rxrec">
+    <None Include="Recordings\BackToHomePage.rxrec">
       <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
       <id>43185cca-9174-42ce-bf52-e4e97de544ea</id>
       <recordtableid>c5b83313-ab86-4e05-9a58-4a98973ad8a0</recordtableid>
     </None>
-    <None Include="Search.rxrec">
+    <None Include="Recordings\CloseBrowser.rxrec">
       <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
-      <id>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</id>
-      <recordtableid>d3aa34cb-271f-4d5e-a490-dfc83757e46c</recordtableid>
+      <id>f2873c10-8693-42cd-a0c8-d76af30a8b60</id>
+      <recordtableid>3407bd7f-4b1b-46f5-a4e7-1c2a187f854f</recordtableid>
     </None>
-    <None Include="OpenBrowser.rxrec">
+    <None Include="Recordings\OpenBrowser.rxrec">
       <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
       <id>716e1a36-fbdd-4a8b-b18e-6b0190f3c31b</id>
       <recordtableid>b2e379b8-fbbe-42d5-bf9a-0983ee661bdf</recordtableid>
     </None>
-    <None Include="Validation.rxrec">
+    <None Include="Recordings\Search.rxrec">
+      <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
+      <id>642b46b7-10cb-4b7e-b0b9-a84c6bb44a50</id>
+      <recordtableid>d3aa34cb-271f-4d5e-a490-dfc83757e46c</recordtableid>
+    </None>
+    <None Include="Recordings\Validation.rxrec">
       <BoundTo>3eab0aee-80cf-4c32-b982-c0d4d93906bd</BoundTo>
       <id>02e26d06-ea6d-4fad-8e56-9ad1339c7022</id>
       <recordtableid>5ab430a7-28be-4e60-b96f-decbbec00186</recordtableid>
@@ -246,6 +246,11 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       <id>05e61672-227d-4c25-bf36-f1ebf60c30dc</id>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Recordings">
+      <id>c3bc6123-701d-4346-84f6-67cc74da957e</id>
+    </Folder>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(RanorexTargetsPath)" Condition="Exists('$(RanorexTargetsPath)')" />
